Refactor CompletionConfig into types folder
