+++ js/packages/teams-ai/src/prompts/PromptManager.ts
@@ -14,8 +14,9 @@ import { MonologueAugmentation, SequenceAugmentation } from '../augmentations';
 import { DataSource } from '../dataSources';
 import { Memory } from '../MemoryFork';
 import { Tokenizer } from '../tokenizers';
+import { CompletionConfig } from '../types/CompletionConfig';
 import { ConversationHistory } from './ConversationHistory';
-import { CompletionConfig, PromptTemplate } from './PromptTemplate';
+import { PromptTemplate } from './PromptTemplate';
 import { DataSourceSection } from './DataSourceSection';
 import { GroupSection } from './GroupSection';
 import { Prompt } from './Prompt';
