+++ js/packages/teams-ai/src/types/index.ts
@@ -1,0 +1,10 @@
+/**
+ * @module teams-ai
+ */
+/**
+ * Copyright (c) Microsoft Corporation. All rights reserved.
+ * Licensed under the MIT License.
+ */
+
+export * from './CompletionConfig';
+export * from './OpenAIFuncion';
